requirements.txt
#requirements
fastapi
uvicorn
pydantic
typing
psycopg2-binary
email-validator
python-multipart
bcrypt